+++ deployment/run-unit-tests.sh
@@ -32,5 +32,12 @@ npm test
 cd ../step-functions
 npm test
 
+# If you're running these commands on macOS and Python3 has been installed using Homebrew, you might see this issue:
+#    DistutilsOptionError: must supply either home or prefix/exec-prefix
+# Please follow the workaround suggested on this StackOverflow answer: https://stackoverflow.com/a/44728772
 cd ../mediainfo
-python3 setup.py test
+rm -rf ./pytests && mkdir ./pytests
+cp lambda_function.py ./test*.py ./pytests
+pip3 install boto3 -t ./pytests
+python3 -m unittest discover -s ./pytests -v
+rm -rf ./pytests