+++ source/custom-resource/lib/cfn/index.spec.js
@@ -1,5 +1,5 @@
 /*********************************************************************************************************************
- *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
+ *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
  *  with the License. A copy of the License is located at                                                             *
@@ -15,33 +15,33 @@ const axios = require('axios');
 const expect = require('chai').expect;
 const MockAdapter = require('axios-mock-adapter');
 
-let lambda = require('./index.js');
-let _event = {
-    RequestType: "Create",
-    ServiceToken: "arn:aws:lambda",
-    ResponseURL: "https://cloudformation",
-    StackId: "arn:aws:cloudformation",
-    RequestId: "63e8ffa2-3059-4607-a450-119d473c73bc",
-    LogicalResourceId: "Uuid",
-    ResourceType: "Custom::UUID",
+const lambda = require('./index.js');
+const _event = {
+    RequestType: 'Create',
+    ServiceToken: 'arn:aws:lambda',
+    ResponseURL: 'https://cloudformation',
+    StackId: 'arn:aws:cloudformation',
+    RequestId: '63e8ffa2-3059-4607-a450-119d473c73bc',
+    LogicalResourceId: 'Uuid',
+    ResourceType: 'Custom::UUID',
     ResourceProperties: {
-        ServiceToken: "arn:aws:lambda",
-        Resource: "abc"
+        ServiceToken: 'arn:aws:lambda',
+        Resource: 'abc'
     }
 };
 
-let _context = {
+const _context = {
     logStreamName: 'cloudwatch'
 };
 
-let _responseStatus = 'ok';
-let _responseData = {
+const _responseStatus = 'ok';
+const _responseData = {
     test: 'testing'
 };
 
 describe('#CFN RESONSE::', () => {
     it('should return "200" on a send cfn response sucess', async () => {
-        let mock = new MockAdapter(axios);
+        const mock = new MockAdapter(axios);
         mock.onPut().reply(200, {});
 
         lambda.send(_event, _context, _responseStatus, _responseData, (err, res) => {
@@ -50,11 +50,11 @@ describe('#CFN RESONSE::', () => {
     });
 
     it('should return "Network Error" on connection timedout', async () => {
-        let mock = new MockAdapter(axios);
+        const mock = new MockAdapter(axios);
         mock.onPut().networkError();
 
         await lambda.send(_event, _context, _responseStatus, _responseData).catch(err => {
-            expect(err.toString()).to.equal("Error: Network Error");
+            expect(err.toString()).to.equal('Error: Network Error');
         });
     });
 });