+++ source/custom-resource/lib/mediapackage/index.spec.js
@@ -1,5 +1,5 @@
 /*********************************************************************************************************************
- *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
+ *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
  *  with the License. A copy of the License is located at                                                             *
@@ -11,24 +11,26 @@
  *  and limitations under the License.                                                                                *
  *********************************************************************************************************************/
 
-'use strict';
 const expect = require('chai').expect;
 const path = require('path');
 const AWS = require('aws-sdk-mock');
 AWS.setSDK(path.resolve('./node_modules/aws-sdk'));
 
 const lambda = require('./index.js');
+const testAssets = require('./test-assets');
 
-const _stackName = 'test';
+const _stackName = 'test-stack';
 const _groupId = 'test-packaging-group';
 
 const validParameters = {
     StackName: _stackName,
     GroupId: _groupId,
-    PackagingConfigurations: 'HLS,DASH'
+    PackagingConfigurations: 'HLS,DASH',
+    DistributionId: testAssets.DistributionId,
+    EnableMediaPackage: 'true'
 };
 
-const assetsResponse = {
+const listAssetsResponse = {
     Assets: [{
         Arn: 'asset-arn',
         Id: 'asset-id',
@@ -39,7 +41,7 @@ const assetsResponse = {
     }]
 };
 
-const configurationsResponse = {
+const listConfigurationsResponse = {
     PackagingConfigurations: [{
         Arn: 'config-arn',
         Id: 'MSS',
@@ -53,13 +55,22 @@ const configurationsResponse = {
     }]
 };
 
+const groupResponse = {
+    Id: _groupId,
+    DomainName: testAssets.DomainName
+};
+
 describe('#MEDIAPACKAGE-VOD::', () => {
-    afterEach(() => AWS.restore('MediaPackageVod'));
+    afterEach(() => {
+        AWS.restore('MediaPackageVod');
+        AWS.restore('CloudFront');
+    });
 
     describe('Create', () => {
         it('should succeed with valid parameters', async () => {
-            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve());
+            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve(groupResponse));
             AWS.mock('MediaPackageVod', 'createPackagingConfiguration', Promise.resolve());
+            AWS.mock('CloudFront', 'getDistributionConfig', Promise.resolve(testAssets.ConfigurationWithMP));
 
             const response = await lambda.create(validParameters);
             expect(response.GroupId).to.equal(_groupId);
@@ -68,16 +79,18 @@ describe('#MEDIAPACKAGE-VOD::', () => {
         it('should ignore duplicate configurations', async () => {
             let callCount = 0;
 
-            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve());
+            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve(groupResponse));
             AWS.mock('MediaPackageVod', 'createPackagingConfiguration', () => {
                 callCount++;
                 return Promise.resolve();
             });
+            AWS.mock('CloudFront', 'getDistributionConfig', Promise.resolve(testAssets.ConfigurationWithMP));
 
             const duplicateConfigParams = {
                 StackName: _stackName,
                 GroupId: _groupId,
-                PackagingConfigurations: 'HLS,DASH,HLS'
+                PackagingConfigurations: 'HLS,DASH,HLS',
+                DistributionId: testAssets.DistributionId
             };
 
             const response = await lambda.create(duplicateConfigParams);
@@ -86,13 +99,15 @@ describe('#MEDIAPACKAGE-VOD::', () => {
         });
 
         it('should succeed when at least one valid configuration is informed', async () => {
-            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve());
+            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve(groupResponse));
             AWS.mock('MediaPackageVod', 'createPackagingConfiguration', Promise.resolve());
+            AWS.mock('CloudFront', 'getDistributionConfig', Promise.resolve(testAssets.ConfigurationWithMP));
 
             const singleInvalidParams = {
                 StackName: _stackName,
                 GroupId: _groupId,
-                PackagingConfigurations: 'HLS,bogus'
+                PackagingConfigurations: 'HLS,bogus',
+                DistributionId: testAssets.DistributionId
             };
 
             const response = await lambda.create(singleInvalidParams);
@@ -114,7 +129,7 @@ describe('#MEDIAPACKAGE-VOD::', () => {
         });
 
         it('should fail when no valid configurations are informed', async () => {
-            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve());
+            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve(groupResponse));
 
             const invalidParameters = {
                 StackName: _stackName,
@@ -133,8 +148,24 @@ describe('#MEDIAPACKAGE-VOD::', () => {
         });
 
         it('should fail when createPackagingConfiguration throws an exception', async () => {
-            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve());
+            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve(groupResponse));
             AWS.mock('MediaPackageVod', 'createPackagingConfiguration', Promise.reject('some error'));
+
+            try {
+                await lambda.create(validParameters);
+            } catch (error) {
+                expect(error).to.not.be.null;
+                expect(error).to.equal('some error');
+                return;
+            }
+
+            expect.fail('exception should have been thrown');
+        });
+
+        it('should fail when getDistributionConfig throws an exception', async () => {
+            AWS.mock('MediaPackageVod', 'createPackagingGroup', Promise.resolve(groupResponse));
+            AWS.mock('MediaPackageVod', 'createPackagingConfiguration', Promise.resolve());
+            AWS.mock('CloudFront', 'getDistributionConfig', Promise.reject('some error'));
 
             try {
                 await lambda.create(validParameters);
@@ -148,10 +179,48 @@ describe('#MEDIAPACKAGE-VOD::', () => {
         });
     });
 
+    describe('Update', () => {
+        it('should not try to add the origin if mediapackage is not enabled', async () => {
+            let callCount = 0;
+
+            AWS.mock('MediaPackageVod', 'describePackagingGroup', Promise.resolve(groupResponse));
+            AWS.mock('CloudFront', 'getDistributionConfig', Promise.resolve(testAssets.ConfigurationWithS3));
+            AWS.mock('CloudFront', 'updateDistribution', () => {
+                callCount++;
+                return Promise.resolve();
+            });
+
+            const disabledMediaPackageParams = {
+                StackName: _stackName,
+                GroupId: _groupId,
+                PackagingConfigurations: 'HLS,DASH',
+                DistributionId: testAssets.DistributionId,
+                EnableMediaPackage: 'false'
+            };
+
+            await lambda.update(disabledMediaPackageParams);
+            expect(callCount).to.equal(0);
+        });
+
+        it('should add the origin if it does not exist', async () => {
+            let callCount = 0;
+
+            AWS.mock('MediaPackageVod', 'describePackagingGroup', Promise.resolve(groupResponse));
+            AWS.mock('CloudFront', 'getDistributionConfig', Promise.resolve(testAssets.ConfigurationWithS3));
+            AWS.mock('CloudFront', 'updateDistribution', () => {
+                callCount++;
+                return Promise.resolve();
+            });
+
+            await lambda.update(validParameters);
+            expect(callCount).to.equal(1);
+        });
+    });
+
     describe('Delete', () => {
         it('should succeed when group exists', async () => {
-            AWS.mock('MediaPackageVod', 'listAssets', Promise.resolve(assetsResponse));
-            AWS.mock('MediaPackageVod', 'listPackagingConfigurations', Promise.resolve(configurationsResponse));
+            AWS.mock('MediaPackageVod', 'listAssets', Promise.resolve(listAssetsResponse));
+            AWS.mock('MediaPackageVod', 'listPackagingConfigurations', Promise.resolve(listConfigurationsResponse));
 
             AWS.mock('MediaPackageVod', 'deleteAsset', Promise.resolve());
             AWS.mock('MediaPackageVod', 'deletePackagingConfiguration', Promise.resolve());
@@ -179,8 +248,8 @@ describe('#MEDIAPACKAGE-VOD::', () => {
         });
 
         it('should fail when exception is unknown', async () => {
-            AWS.mock('MediaPackageVod', 'listAssets', Promise.resolve(assetsResponse));
-            AWS.mock('MediaPackageVod', 'listPackagingConfigurations', Promise.resolve(configurationsResponse));
+            AWS.mock('MediaPackageVod', 'listAssets', Promise.resolve(listAssetsResponse));
+            AWS.mock('MediaPackageVod', 'listPackagingConfigurations', Promise.resolve(listConfigurationsResponse));
 
             AWS.mock('MediaPackageVod', 'deleteAsset', Promise.resolve());
             AWS.mock('MediaPackageVod', 'deletePackagingConfiguration', Promise.resolve());