+++ source/dynamo/lib/error.js
@@ -1,5 +1,5 @@
 /*********************************************************************************************************************
- *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
+ *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
  *  with the License. A copy of the License is located at                                                             *
@@ -20,10 +20,10 @@ let errHandler = async (event, _err) => {
 
     try {
         let payload = {
-            "guid": event.guid,
-            "event": event,
-            "function": process.env.AWS_LAMBDA_FUNCTION_NAME,
-            "error": _err.toString()
+            'guid': event.guid,
+            'event': event,
+            'function': process.env.AWS_LAMBDA_FUNCTION_NAME,
+            'error': _err.toString()
         };
 
         let params = {