+++ source/input-validate/lib/index.spec.js
@@ -1,5 +1,5 @@
 /*********************************************************************************************************************
- *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
+ *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
  *  with the License. A copy of the License is located at                                                             *
@@ -11,12 +11,12 @@
  *  and limitations under the License.                                                                                *
  *********************************************************************************************************************/
 
-let expect = require('chai').expect;
-var path = require('path');
-let AWS = require('aws-sdk-mock');
+const expect = require('chai').expect;
+const path = require('path');
+const AWS = require('aws-sdk-mock');
 AWS.setSDK(path.resolve('./node_modules/aws-sdk'));
 
-let lambda = require('../index.js');
+const lambda = require('../index.js');
 
 describe('#INPUT VALIDATE::', () => {
     process.env.ErrorHandler = 'error_handler';
@@ -25,8 +25,11 @@ describe('#INPUT VALIDATE::', () => {
     process.env.Source = 'source_bucket';
     process.env.EnableMediaPackage = 'true';
     process.env.InputRotate = 'DEGREE_0';
+    process.env.AcceleratedTranscoding = 'DISABLED'
+    process.env.EnableSns = 'true';
+    process.env.EnableSqs = 'true';
 
-    let _video = {
+    const _video = {
         workflowTrigger: 'Video',
         guid: '1234-1223232-212121',
         Records: [{
@@ -38,7 +41,7 @@ describe('#INPUT VALIDATE::', () => {
         }]
     };
 
-    let _json = {
+    const _json = {
         workflowTrigger: 'Metadata',
         guid: '1234-1223232-212121',
         Records: [{
@@ -50,24 +53,22 @@ describe('#INPUT VALIDATE::', () => {
         }]
     };
 
-    afterEach(() => {
-        AWS.restore('S3');
-    });
+    afterEach(() => AWS.restore('S3'));
 
-    it('should return "SUCCESS" when validating souce video', async () => {
-        let response = await lambda.handler(_video);
+    it('should succeed when processing valid source video', async () => {
+        const response = await lambda.handler(_video);
         expect(response.srcVideo).to.equal('video.mp4');
     });
 
-    it('should return "SUCCESS" when validating metadata', async () => {
-        let validMetadata = {
+    it('should succeed when processing valid metadata', async () => {
+        const validMetadata = {
             "Body": '{"srcVideo": "video_from_json.mp4", "archiveSource": false, "frameCapture": false, "srcBucket": "other-source", "jobTemplate_720p": "other-template"}'
         };
 
         AWS.mock('S3', 'getObject', Promise.resolve(validMetadata));
         AWS.mock('S3', 'headObject', Promise.resolve());
 
-        let response = await lambda.handler(_json);
+        const response = await lambda.handler(_json);
         expect(response.srcVideo).to.equal('video_from_json.mp4');
         expect(response.archiveSource).to.be.false;
         expect(response.frameCapture).to.be.false;
@@ -78,26 +79,26 @@ describe('#INPUT VALIDATE::', () => {
     });
 
     it('should always use MediaPackage env variable', async () => {
-        let metadata = {
+        const metadata = {
             "Body": '{"srcVideo": "video_from_json.mp4", "enableMediaPackage": false }'
         };
 
         AWS.mock('S3', 'getObject', Promise.resolve(metadata));
         AWS.mock('S3', 'headObject', Promise.resolve());
 
-        let response = await lambda.handler(_json);
+        const response = await lambda.handler(_json);
         expect(response.enableMediaPackage).to.be.true;
     });
 
     it('should correctly handle metadata in PascalCase', async () => {
-        let invalidMetadata = {
+        const invalidMetadata = {
             "Body": '{"srcVideo": "video_from_json.mp4", "ArchiveSource": false, "FrameCapture": false, "SrcBucket": "other-source", "inputRotate": "AUTO" }'
         };
 
         AWS.mock('S3', 'getObject', Promise.resolve(invalidMetadata));
         AWS.mock('S3', 'headObject', Promise.resolve());
 
-        let response = await lambda.handler(_json);
+        const response = await lambda.handler(_json);
         expect(response.srcVideo).to.equal('video_from_json.mp4');
         expect(response.archiveSource).to.be.false;
         expect(response.frameCapture).to.be.false;
@@ -105,13 +106,11 @@ describe('#INPUT VALIDATE::', () => {
         expect(response.inputRotate).to.equal('AUTO');
     });
 
-    it('should return "S3 GET ERROR" when validating metadata', async () => {
+    it('should fail when getting object from S3 throws an exception', async () => {
         AWS.mock('S3', 'getObject', Promise.reject('S3 GET ERROR'));
         AWS.mock('S3', 'headObject', Promise.resolve());
         AWS.mock('Lambda', 'invoke', Promise.resolve());
 
-        await lambda.handler(_json).catch(err => {
-            expect(err).to.equal('S3 GET ERROR');
-        });
+        await lambda.handler(_json).catch(err => expect(err).to.equal('S3 GET ERROR'));
     });
 });