+++ source/sns-notification/lib/index.spec.js
@@ -1,5 +1,5 @@
 /*********************************************************************************************************************
- *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
+ *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
  *  with the License. A copy of the License is located at                                                             *
@@ -31,7 +31,7 @@ describe('#SNS::', () => {
         startTime: 'now',
         srcVideo: 'video.mp4',
         workflowStatus: 'Complete',
-        jobTemplate_2160p: 'jobTemplate_720p',
+        jobTemplate_720p: 'jobTemplate_720p',
         jobTemplate_1080p: 'jobTemplate_1080p',
         jobTemplate_2160p: 'jobTemplate_1080p'
     };