+++ source/sqs-publish/lib/index.spec.js
@@ -1,5 +1,5 @@
 /*********************************************************************************************************************
- *  Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
+ *  Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.                                           *
  *                                                                                                                    *
  *  Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance    *
  *  with the License. A copy of the License is located at                                                             *
@@ -11,35 +11,38 @@
  *  and limitations under the License.                                                                                *
  *********************************************************************************************************************/
 
-const AWS = require('aws-sdk');
+const expect = require('chai').expect;
+const path = require('path');
+const AWS = require('aws-sdk-mock');
+AWS.setSDK(path.resolve('./node_modules/aws-sdk'));
 
-let errHandler = async (event, _err) => {
-    const lambda = new AWS.Lambda({
-        region: process.env.AWS_REGION
+const lambda = require('../index.js');
+
+describe('#SNS::', () => {
+    const _event = {
+        guid: '12345678',
+        startTime: 'now',
+        srcVideo: 'video.mp4',
+    };
+
+    process.env.ErrorHandler = 'error_handler';
+    process.env.SqsQueue = 'https://sqs.amazonaws.com/1234'
+
+    afterEach(() => AWS.restore('SQS'));
+
+    it('should return "success" on send SQS message', async () => {
+        AWS.mock('SQS', 'sendMessage', Promise.resolve());
+
+        const response = await lambda.handler(_event);
+        expect(response.srcVideo).to.equal('video.mp4');
     });
 
-    try {
-        let payload = {
-            'guid': event.guid,
-            'event': event,
-            'function': process.env.AWS_LAMBDA_FUNCTION_NAME,
-            'error': _err.toString()
-        };
+    it('should return "SQS ERROR" when send SQS message fails', async () => {
+        AWS.mock('SQS', 'sendMessage', Promise.reject('SQS ERROR'));
+        AWS.mock('Lambda', 'invoke', Promise.resolve());
 
-        let params = {
-            FunctionName: process.env.ErrorHandler,
-            Payload: JSON.stringify(payload, null, 2)
-        };
-
-        await lambda.invoke(params).promise();
-    } catch (err) {
-        console.log(err);
-        throw err;
-    }
-
-    return 'success';
-};
-
-module.exports = {
-    handler: errHandler
-};
+        await lambda.handler(_event).catch(err => {
+            expect(err).to.equal('SQS ERROR');
+        });
+    });
+});