+++ CHANGELOG.md
@@ -4,6 +4,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [5.1.0] - 2020-04-30
+### Added
+- Default encryption to SNS topic
+- Environment variable to configure the AWS SDK to reuse TCP connections (https://docs.aws.amazon.com/sdk-for-javascript/v2/developer-guide/node-reusing-connections.html)
+- SQS queue options at deployment to capture workflow outputs
+- Support for accelerated transcoding in Elemental MediaCoonvert: https://aws.amazon.com/about-aws/whats-new/2019/10/announcing-new-aws-elemental-mediaconvert-features-for-accelerated-transcoding-dash-and-avc-video-quality/
+- support for Glacier deep archive
+
+### Changed
+- Lambda functions runtime to latest available (Node.js 12)
+- Build assets to include package-lock.json files
+- Build and test commands to use _npm ci_ instead of _npm install_
+- Cloudformation template to use _AWS::Partition_ instead of _aws_
+- Logic to add MediaPackage VOD as a custom origin to CloudFront (it's now done as a custom resource when the stack is created / updated)
+
+### Fixed
+- Links in README file
+- fix to buildUrl function in output-validate lambda to support non root objects (https://github.com/awslabs/video-on-demand-on-aws/issues/61)
+- fix mediainfo lambda function signing method error (https://github.com/awslabs/video-on-demand-on-aws/issues/670
+
+### Removed
+- _'use strict'_ directives
+
 ## [5.0.0] - 2019-11-20
 ### Added
 - MediaPackage VOD support
@@ -35,4 +59,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lambda functions (except for mediainfo) runtime to nodejs10.x
 - Mediainfo lambda function to python3.7
 - Mediainfo executable version (from v0.7.92.1 to v19.09)
-